[refactor] extract reducer function
